refactor: rename fileExplorer
