+++ ui/src/components/inputs/FileExplorerWrapper.vue
@@ -1,5 +1,5 @@
 <template>
-    <EditorSidebar
+    <FileExplorer
         :currentNS="namespace"
         style="width: 100%;height: 100%;"
     />
@@ -7,7 +7,7 @@
 
 <script setup lang="ts">
     import {computed} from "vue";
-    import EditorSidebar from "./EditorSidebar.vue";
+    import FileExplorer from "./FileExplorer.vue";
     import {useFlowStore} from "../../stores/flow";
 
     const flowStore = useFlowStore();
